+++ src/webgl1/polyfill-context.js
@@ -8,11 +8,16 @@
 // easy to reuse or repurpose in other projects.
 
 /* eslint-disable camelcase, brace-style */
+import GL from '../constants';
 import {getParameterPolyfill} from './polyfill-get-parameter';
 import polyfillVertexArrayObject from './polyfill-vertex-array-object';
-import {WebGLRenderingContext} from '../webgl-utils';
-import assert from '../utils/assert';
-import GL from '../constants';
+import {WebGLRenderingContext} from './webgl-rendering-context';
+
+function assert(condition, message) {
+  if (!condition) {
+    throw new Error(message || 'luma.gl: assertion failed.');
+  }
+}
 
 const OES_vertex_array_object = 'OES_vertex_array_object';
 const ANGLE_instanced_arrays = 'ANGLE_instanced_arrays';
@@ -234,3 +239,7 @@ export default function polyfillContext(gl) {
   }
   return gl;
 }
+
+/* global window, global */
+const global_ = typeof global !== 'undefined' ? global : window;
+global_.polyfillContext = polyfillContext;