+++ test/src/webgl1/index.js
@@ -1,0 +1,2 @@
+import './polyfill-context.spec';
+import './polyfill-get-parameter.spec';