+++ webgl1.js
@@ -1,0 +1,2 @@
+// Installs polyfills for webgl1
+require('./dist/es5/webgl1');